feat: add table for team members and return their data as part of the Event data structure
