+++ app/Models/TeamMemberModel.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Models;
+
+class TeamMemberModel extends GenericRoleModel
+{
+    protected $table = 'TeamMember';
+}
